Implemented test and validation for json payloads created from netCDF files
