+++ examples/air_temperature_gullingen_skisenter-parent_meta.json
@@ -1,0 +1,44 @@
+{
+    "type": "Feature",
+    "geometry": {
+        "type": "Point",
+        "coordinates": [
+            59.4128,
+            6.4697
+        ]
+    },
+    "properties": {
+        "title": "Air temperature observations from weather station Gullingen , ROGALAND, Norge, (WIGOS ID: 0-578-0-46220, station ID: 46220).",
+        "data_id": "8a0744b7-f0dc-4929-8700-f1285e76df36",
+        "metadata_id": "no.met:8a0744b7-f0dc-4929-8700-f1285e76df36",
+        "keywords": "GCMDSK:EARTH SCIENCE > ATMOSPHERE > ATMOSPHERIC TEMPERATURE > SURFACE TEMPERATURE > AIR TEMPERATURE, GCMDLOC:Geographic Region > Northern Hemisphere, GCMDPROV: Government Agencies-non-US > Norway > NO/MET > Norwegian Meteorological Institute, GEMET:Meteorological geographical features, GEMET:Atmospheric conditions, GEMET:Oceanographic geographical features, NORTHEMES:Weather and climate",
+        "Conventions": "CF-1.10, ACDD-1.3",
+        "history": "2023-06-14T14:38:30.363447+00:00: Created from Frost.",
+        "summary": "Time series of Air temperature observations from the Norwegian weather station Gullingen  (WIGOS ID 0-578-0-46220, station ID 46220). The observations have been through the data collection system of the Norwegian Meteorological institute which includes a number of automated and manual quality control routines. The number of available quality control routines is element dependent.",
+        "institution": "Norwegian Meteorological Institute (MET Norway)",
+        "source": "In Situ Land-based station",
+        "creator_name": "Norwegian Meteorological Institute",
+        "creator_url": "https://www.met.no",
+        "creator_email": "observation_data_archive@met.no",
+        "license": "http://spdx.org/licenses/CC-BY-4.0(CC-BY-4.0)",
+        "access_constraint": "Open",
+        "datetime": "2023-06-14T14:00:00.000000000",
+        "pubtime": "2023-08-22T07:16:37.000000000Z"
+    },
+    "links": [
+        {
+            "href": "https://data.met.no/dataset/8a0744b7-f0dc-4929-8700-f1285e76df36(Dataset landing page)",
+            "rel": "item",
+            "type": "application/json"
+        }
+    ],
+    "version": "v0.1",
+    "content": {
+        "encoding": "utf-8",
+        "standard_name": "air_temperature",
+        "unit": "degC",
+        "size": 18,
+        "value": "22.899999618530273"
+    },
+    "id": "dddf0727-7f05-4a1f-a56a-73a6f0d0fceb"
+}
