+++ schemas/e-soh-message-spec.json
@@ -1,8 +1,7 @@
 {
-  "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "$id": "https://schemas.wmo.int/wis2/broker/message/0.9.0/schema.yml",
-  "title": "WMO WIS 2.0 broker message schema",
-  "description": "WMO WIS 2.0 broker message schema",
+  "$schema": "https://json-schema.org/draft/2020-12/schema#",
+  "title": "E-SOH broker message schema",
+  "description": "E-SOH broker message schema",
   "type": "object",
   "properties": {
     "id": {
@@ -13,7 +12,7 @@
     "version": {
       "type": "string",
       "description": "Version of message specification.",
-      "const": "v04"
+      "const": "v0.1"
     },
     "type": {
       "type": "string",
@@ -102,10 +101,7 @@
           "description": "A paragraph describing the dataset, analogous to an abstract for a paper."
         },
         "keywords": {
-          "type": "array",
-          "items": {
-            "type": "string"
-          },
+          "type": "string",
           "description": "A comma-separated list of key words and/or phrases. Keywords may be common words or phrases, terms from a controlled vocabulary (GCMD is often used), or URIs for terms from a controlled vocabulary (see also 'keywords_vocabulary' attribute)."
         },
         "license": {
@@ -113,10 +109,7 @@
           "description": "Provide the URL to a standard or specific license, enter 'Freely Distributed' or 'None', or describe any restrictions to data access and distribution in free text."
         },
         "Conventions": {
-          "type": "array",
-          "items": {
-            "type": "string"
-          },
+          "type": "string",
           "description": "A comma-separated list of the conventions that are followed by the dataset. For files that follow this version of ACDD, include the string 'ACDD-1.3'. (This attribute is described in the NetCDF Users Guide.)"
         },
         "naming_authority": {
@@ -124,7 +117,7 @@
           "description": "The organization that provides the initial id (see above) for the dataset. The naming authority should be uniquely specified by this attribute. We recommend using reverse-DNS naming for the naming authority; URIs are also acceptable. Example: 'edu.ucar.unidata'."
         },
         "history": {
-          "type": "array",
+          "type": "string",
           "description": "Provides an audit trail for modifications to the original data. This attribute is also in the NetCDF Users Guide: 'This is a character array with a line for each invocation of a program that has modified the dataset. Well-behaved generic netCDF applications should append a line containing: date, time of day, user name, program name and command arguments.' To include a more complete description you can append a reference to an ISO Lineage entity; see NOAA EDM ISO Lineage guidance."
         },
         "metadata_id": {
@@ -336,44 +329,5 @@
     "geometry",
     "properties",
     "links"
-  ],
-  "example": {
-    "id": "31e9d66a-cd83-4174-9429-b932f1abe1be",
-    "version": "v04",
-    "type": "Feature",
-    "geometry": {
-      "type": "Point",
-      "coordinates": [
-        6.146255135536194,
-        46.223296618227444
-      ]
-    },
-    "properties": {
-      "pubtime": "2022-03-20T04:50:18.314854383Z",
-      "datetime": "2022-03-20T04:45:00Z",
-      "integrity": {
-        "method": "sha512",
-        "value": "A2KNxvks...S8qfSCw=="
-      },
-      "data_id": "dataset/123/data-granule/UANT01_CWAO_200445___15103.bufr4",
-      "metadata_id": "urn:x-wmo:md:can:eccc-msc:observations.swob",
-      "content": {
-        "encoding": "utf-8",
-        "value": "encoded bytes from the file",
-        "size": 457
-      }
-    },
-    "links": [
-      {
-        "href": "https://example.org/data/4Pubsub/92c557ef-d28e-4713-91af-2e2e7be6f8ab.bufr4",
-        "rel": "canonical",
-        "type": "application/x-bufr"
-      },
-      {
-        "href": "https://example.org/oapi/collections/my-dataset/items/my-data-granule",
-        "rel": "item",
-        "type": "application/json"
-      }
-    ]
-  }
+  ]
 }
